requirements.txt
numpy
matplotlib
opencv-python
pybullet
torch
easyocr
scipy
kornia
scikit-learn
scikit-image
pyspellchecker
Pillow
